Initial Commit-A405FNXXU3CUC2

+++ drivers/media/rc/imon.c
@@ -573,8 +573,6 @@ static int send_packet(struct imon_context *ictx)
 	int retval = 0;
 	struct usb_ctrlrequest *control_req = NULL;
 
-	lockdep_assert_held(&ictx->lock);
-
 	/* Check if we need to use control or interrupt urb */
 	if (!ictx->tx_control) {
 		pipe = usb_sndintpipe(ictx->usbdev_intf0,
@@ -623,18 +621,16 @@ static int send_packet(struct imon_context *ictx)
 		smp_rmb(); /* ensure later readers know we're not busy */
 		pr_err_ratelimited("error submitting urb(%d)\n", retval);
 	} else {
-		/* Wait for transmission to complete (or abort or timeout) */
-		retval = wait_for_completion_interruptible_timeout(&ictx->tx.finished, 10 * HZ);
-		if (retval <= 0) {
+		/* Wait for transmission to complete (or abort) */
+		mutex_unlock(&ictx->lock);
+		retval = wait_for_completion_interruptible(
+				&ictx->tx.finished);
+		if (retval) {
 			usb_kill_urb(ictx->tx_urb);
 			pr_err_ratelimited("task interrupted\n");
-			if (retval < 0)
-				ictx->tx.status = retval;
-			else
-				ictx->tx.status = -ETIMEDOUT;
 		}
+		mutex_lock(&ictx->lock);
 
-		ictx->tx.busy = false;
 		retval = ictx->tx.status;
 		if (retval)
 			pr_err_ratelimited("packet tx failed (%d)\n", retval);
@@ -943,8 +939,7 @@ static ssize_t vfd_write(struct file *file, const char __user *buf,
 		return -ENODEV;
 	}
 
-	if (mutex_lock_interruptible(&ictx->lock))
-		return -ERESTARTSYS;
+	mutex_lock(&ictx->lock);
 
 	if (!ictx->dev_present_intf0) {
 		pr_err_ratelimited("no iMON device present\n");
@@ -1115,7 +1110,7 @@ static int imon_ir_change_protocol(struct rc_dev *rc, u64 *rc_type)
 	int retval;
 	struct imon_context *ictx = rc->priv;
 	struct device *dev = ictx->dev;
-	const bool unlock = mutex_trylock(&ictx->lock);
+	bool unlock = false;
 	unsigned char ir_proto_packet[] = {
 		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x86 };
 
@@ -1143,6 +1138,11 @@ static int imon_ir_change_protocol(struct rc_dev *rc, u64 *rc_type)
 	}
 
 	memcpy(ictx->usb_tx_buf, &ir_proto_packet, sizeof(ir_proto_packet));
+
+	if (!mutex_is_locked(&ictx->lock)) {
+		unlock = true;
+		mutex_lock(&ictx->lock);
+	}
 
 	retval = send_packet(ictx);
 	if (retval)
@@ -1644,7 +1644,8 @@ static void imon_incoming_packet(struct imon_context *ictx,
 	spin_unlock_irqrestore(&ictx->kc_lock, flags);
 
 	/* send touchscreen events through input subsystem if touchpad data */
-	if (ictx->touch && len == 8 && buf[7] == 0x86) {
+	if (ictx->display_type == IMON_DISPLAY_TYPE_VGA && len == 8 &&
+	    buf[7] == 0x86) {
 		imon_touch_event(ictx, buf);
 		return;
 
@@ -1755,6 +1756,14 @@ static void usb_rx_callback_intf0(struct urb *urb)
 	if (!ictx)
 		return;
 
+	/*
+	 * if we get a callback before we're done configuring the hardware, we
+	 * can't yet process the data, as there's nowhere to send it, but we
+	 * still need to submit a new rx URB to avoid wedging the hardware
+	 */
+	if (!ictx->dev_present_intf0)
+		goto out;
+
 	switch (urb->status) {
 	case -ENOENT:		/* usbcore unlink successful! */
 		return;
@@ -1763,22 +1772,8 @@ static void usb_rx_callback_intf0(struct urb *urb)
 		break;
 
 	case 0:
-		/*
-		 * if we get a callback before we're done configuring the hardware, we
-		 * can't yet process the data, as there's nowhere to send it, but we
-		 * still need to submit a new rx URB to avoid wedging the hardware
-		 */
-		if (ictx->dev_present_intf0)
-			imon_incoming_packet(ictx, urb, intfnum);
+		imon_incoming_packet(ictx, urb, intfnum);
 		break;
-
-	case -ECONNRESET:
-	case -EILSEQ:
-	case -EPROTO:
-	case -EPIPE:
-		dev_warn(ictx->dev, "imon %s: status(%d)\n",
-			 __func__, urb->status);
-		return;
 
 	default:
 		dev_warn(ictx->dev, "imon %s: status(%d): ignored\n",
@@ -1786,6 +1781,7 @@ static void usb_rx_callback_intf0(struct urb *urb)
 		break;
 	}
 
+out:
 	usb_submit_urb(ictx->rx_urb_intf0, GFP_ATOMIC);
 }
 
@@ -1801,6 +1797,14 @@ static void usb_rx_callback_intf1(struct urb *urb)
 	if (!ictx)
 		return;
 
+	/*
+	 * if we get a callback before we're done configuring the hardware, we
+	 * can't yet process the data, as there's nowhere to send it, but we
+	 * still need to submit a new rx URB to avoid wedging the hardware
+	 */
+	if (!ictx->dev_present_intf1)
+		goto out;
+
 	switch (urb->status) {
 	case -ENOENT:		/* usbcore unlink successful! */
 		return;
@@ -1809,22 +1813,8 @@ static void usb_rx_callback_intf1(struct urb *urb)
 		break;
 
 	case 0:
-		/*
-		 * if we get a callback before we're done configuring the hardware, we
-		 * can't yet process the data, as there's nowhere to send it, but we
-		 * still need to submit a new rx URB to avoid wedging the hardware
-		 */
-		if (ictx->dev_present_intf1)
-			imon_incoming_packet(ictx, urb, intfnum);
+		imon_incoming_packet(ictx, urb, intfnum);
 		break;
-
-	case -ECONNRESET:
-	case -EILSEQ:
-	case -EPROTO:
-	case -EPIPE:
-		dev_warn(ictx->dev, "imon %s: status(%d)\n",
-			 __func__, urb->status);
-		return;
 
 	default:
 		dev_warn(ictx->dev, "imon %s: status(%d): ignored\n",
@@ -1832,6 +1822,7 @@ static void usb_rx_callback_intf1(struct urb *urb)
 		break;
 	}
 
+out:
 	usb_submit_urb(ictx->rx_urb_intf1, GFP_ATOMIC);
 }
